requirements.txt
annoy==1.17.3
catboost==1.2.8
category_encoders==2.8.1
gensim
imblearn==0.0
ipython==8.12.3
joblib==1.4.2
lightgbm==4.6.0
matplotlib==3.10.1
numpy
optuna==4.2.1
pandas
plotly==6.0.1
psutil==7.0.0
pyarrow==19.0.1
python-dotenv==1.1.0
scikit_learn==1.6.1
seaborn==0.13.2
torch==2.6.0
tqdm==4.67.1
transformers==4.51.2
xgboost==3.0.0
